moved _take_sources and _page_numbers_to_str to VectorDatabase
