+++ ragna/source_storages/_chroma.py
@@ -61,7 +61,7 @@ class Chroma(VectorDatabaseSourceStorage):
             metadatas.append(
                 {
                     "document_id": str(embedding.chunk.document_id),
-                    "page_numbers": EmbeddingModel._page_numbers_to_str(embedding.chunk.page_numbers),
+                    "page_numbers": self._page_numbers_to_str(embedding.chunk.page_numbers),
                     "num_tokens": embedding.chunk.num_tokens,
                 }
             )
@@ -128,7 +128,7 @@ class Chroma(VectorDatabaseSourceStorage):
         #  Thus, we likely need to have a callable parameter for this class
 
         document_map = {str(document.id): document for document in documents}
-        return EmbeddingModel._take_sources_up_to_max_tokens(
+        return self._take_sources_up_to_max_tokens(
             (
                 Source(
                     id=result["id"],
