+++ ragna/source_storages/_vector_database.py
@@ -1,7 +1,56 @@
+from typing import Optional, Iterable, cast
+import itertools
+from ragna._compat import itertools_pairwise
+
+import abc
+
 from ragna.core import (
+    Document,
+    Source,
     SourceStorage,
 )
 
 
 class VectorDatabaseSourceStorage(SourceStorage):
-    ...
+    @classmethod
+    def _page_numbers_to_str(cls, page_numbers: Optional[Iterable[int]]) -> str:
+        if not page_numbers:
+            return ""
+
+        page_numbers = sorted(set(page_numbers))
+        if len(page_numbers) == 1:
+            return str(page_numbers[0])
+
+        ranges_str = []
+        range_int = []
+        for current_page_number, next_page_number in itertools_pairwise(
+            itertools.chain(sorted(page_numbers), [None])
+        ):
+            current_page_number = cast(int, current_page_number)
+
+            range_int.append(current_page_number)
+            if next_page_number is None or next_page_number > current_page_number + 1:
+                ranges_str.append(
+                    ", ".join(map(str, range_int))
+                    if len(range_int) < 3
+                    else f"{range_int[0]}-{range_int[-1]}"
+                )
+                range_int = []
+
+        return ", ".join(ranges_str)
+
+    @classmethod
+    def _take_sources_up_to_max_tokens(
+        cls, sources: Iterable[Source], *, max_tokens: int
+    ) -> list[Source]:
+        taken_sources = []
+        total = 0
+        for source in sources:
+            new_total = total + source.num_tokens
+            if new_total > max_tokens:
+                break
+
+            taken_sources.append(source)
+            total = new_total
+
+        return taken_sources
